Updated version to 0.1.1
Updated dependency on Util to 0.2.13
Allowed the generation of an assembled wiget without specifying the
value from parameter. In this case the assembled wiget gets its values
from it parent assembled wiget in the assembly
Added handling of null data source to AssembledWiget
Added handling of null parameter to AssembledWiget.get(...)
Added AssembledWiget.add(WigetContainer, Class extends Wiget) that
generates a new assembled wiget in the container of this assembled wiget
for the given wiget class and gaining its values from this wiget.
Added handling for the wiget static field not existing when generating
the wiget model. This makes the assmbly of wigets less prone to fail and
adds a warning to the log if the static wiget model is not as expected.
Added WigetAssembly.assemble(Class extends Wiget) to gererate an
assembly of the required wiget within the existing wiget assembly.